Improvements to HBOS merge strategy and testing Added ADOutlierHbos method to set the max number of bins HbosParam update and update_and_return now use a common function to perform the update
Replaced Scott method in Hbos merge and histogram generation with a fixed number of bins based on testing. As the range will always either increase or stay the same, this guarantees the histogram will only ever become coarser and we don't need to try to generate fine bins from coarse bins where the information has already been lost.

When generating a histogram from a dataset with all values the same, the edges are now placed much more closely around that value to avoid introducing artifical discretization errors

Added a benchmark_suite test allowing the histogram generation and merge accuracy to be explored using the KL divergence against the true underlying distribution as a metric