Create a new image for end-to-end tests of CV2.
This commit creates a new image, with corresponding support scripts, to
run the end-to-end tests for the chimbuko-visualization2 (CV2) package.
A README is included to explain its usage.

This commit also updates some of the spack package dependencies for the
CV2 package. There seem to be more incompatibilities that will need to
be resolved in order to fully upgrade CV2.